requirements.txt
pypdf2==3.0.1
python-decouple==3.4

